(fix)Implemented all features
Was having an issue where running the script would cause the terminal
window to move itself behind the window that previously held focus, and
would also produce no output, or errors, and had very garbled input when
reading files.

Was probably an issue with pipenv, but switched away from invoke
already, to argparse, and it's already implemented.